pyproject.toml
[project]
name = "base-module"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.117.1",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
    "sqlmodel>=0.0.25",
]
